Fix #4216: move ast to madx
