+++ sirepo/template/madx.py
@@ -798,8 +798,27 @@ def _format_field_value(state, model, field, el_type):
 
 
 def _format_rpn_value(value):
-    return code_variable.PurePythonEval.postfix_to_infix(value
-    )
+    import astunparse
+    import ast
+    class Visitor(ast.NodeTransformer):
+        def visit_Call(self, node):
+            if node.func.id == 'pow':
+                return ast.BinOp(
+                    left=node.args[0],
+                    op=ast.Pow(),
+                    right=node.args[1],
+                    keywords=[]
+                )
+            return node
+
+    r = code_variable.PurePythonEval.postfix_to_infix(value)
+    if type(r) == str and 'pow' in r:
+        tree = ast.parse(r)
+        for n in ast.walk(tree):
+            Visitor().visit(n)
+            ast.fix_missing_locations(n)
+        r = astunparse.unparse(tree).strip().replace('**', '^')
+    return r
 
 def _generate_commands(filename_map, util):
     _update_beam_energy(util.data)
