+++ sirepo/template/madx_converter.py
@@ -43,7 +43,8 @@ class MadxConverter():
         self.__init_direction(data, self.sim_type, madx.SIM_TYPE)
         self.field_map = self.full_field_map.to_madx
         self.drift_type = self.full_field_map.from_madx.DRIFT[0]
-        return self._convert(data)
+        d = self._convert(data)
+        return d
 
     def to_madx_text(self, data):
         from sirepo.template import madx
