+++ tests/template/madx_converter_data/test1.madx
@@ -4,7 +4,7 @@ REAL lt = 9.25;
 REAL sr_alpha_x = -1.0;
 REAL sr_alpha_y = -0.0;
 REAL sr_gamma = 1.01;
-REAL sr_beta = sqrt(1 - (1 / (sr_gamma ^ 2)));
+REAL sr_beta = sqrt((1 - (1 / (sr_gamma ^ 2))));
 REAL sr_beta_x = 1.1;
 REAL sr_beta_y = 0.7;
 REAL sr_dp_s_coupling = -0.09950371902099893;
